add api tests to check routes for public and private instances of API
